refactor(context-engine): extract swap-controller helpers to keep file under ratchet
context-engine-swap.ts crossed the 400-line soft limit by 75 after the
fail-loud delivery + per-turn attribution work, pushing the repo's
complexity ratchet from 728 to 729. CI's check:complexity blocks on
that single-violation regression.

Move pure helpers (sameIdentity, findFrameMatching, freezeIdentity,
freezeEvent) and the SwapControllerOptions interface into a sibling
context-engine-swap-internals.ts. The split is mechanical — no
behavior change — and the helper file is internal: not re-exported
from src/index.ts. Re-export of SwapControllerOptions is preserved so
external imports stay stable.

Result: swap.ts at 399 lines, ratchet back to 728.
